AgroBot-Universal-AI-Based-Agricultural-Assistant/pyproject.toml
[project]
name = "agrobot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-socketio>=5.6.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.66.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
]
